requirements.txt
# requirements.txt — Mac Apple Silicon (M1/M2/M3/M4)

# Core PyTorch (tự detect MPS, không cần flag CUDA)
torch>=2.2.0
transformers>=4.44.0
accelerate>=0.27.0
sentencepiece>=0.1.99

# MLX — nhanh hơn MPS, khuyên dùng trên M-series
mlx>=0.12.0
mlx-lm>=0.12.0

# Fine-tuning (không dùng bitsandbytes — chưa support MPS)
peft>=0.10.0
trl>=0.8.0
datasets>=2.18.0

# Language detection
langdetect>=1.0.9

# UI
gradio>=4.36.0

# Environment variables
python-dotenv>=1.0.0